@@ -3,7 +3,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 
 class TimestampSerializerPlugin implements SerializerPlugin {
   @override
-  Object beforeSerialize(Object object, FullType specifiedType) {
+  Object? beforeSerialize(Object? object, FullType specifiedType) {
     if (object is DateTime && specifiedType.root == DateTime) {
       return object.toUtc();
     }
@@ -11,15 +11,15 @@ class TimestampSerializerPlugin implements SerializerPlugin {
   }
 
   @override
-  Object afterSerialize(Object object, FullType specifiedType) {
+  Object? afterSerialize(Object? object, FullType specifiedType) {
     if (specifiedType.root == DateTime) {
-      return Timestamp.fromMicrosecondsSinceEpoch(object);
+      return Timestamp.fromMicrosecondsSinceEpoch(object as int);
     }
     return object;
   }
 
   @override
-  Object beforeDeserialize(Object object, FullType specifiedType) {
+  Object? beforeDeserialize(Object? object, FullType specifiedType) {
     if (object is Timestamp && specifiedType.root == DateTime) {
       return object.microsecondsSinceEpoch;
       // return object.toDate();
@@ -29,7 +29,7 @@ class TimestampSerializerPlugin implements SerializerPlugin {
   }
 
   @override
-  Object afterDeserialize(Object object, FullType specifiedType) {
+  Object? afterDeserialize(Object? object, FullType specifiedType) {
     return object;
   }
 }
